fixup! feat(page-services): refactor status & running_status usage
