+++ libs/pages/services/src/lib/ui/page-general/page-general.spec.tsx
@@ -1,13 +1,24 @@
 import { render } from '__tests__/utils/setup-jest'
+import { ServiceDeploymentStatusEnum, StateEnum } from 'qovery-typescript-axios'
 import { applicationFactoryMock } from '@qovery/shared/factories'
 import { PageGeneral, PageGeneralProps } from './page-general'
 
 let props: PageGeneralProps
 
+const applications = applicationFactoryMock(2)
+
 beforeEach(() => {
   props = {
     environmentMode: '',
-    services: applicationFactoryMock(2),
+    services: applications.map(({ id, ...rest }) => ({
+      id,
+      ...rest,
+      status: {
+        state: StateEnum.STOPPED,
+        id,
+        service_deployment_status: ServiceDeploymentStatusEnum.UP_TO_DATE,
+      },
+    })),
     buttonActions: [
       {
         name: 'stop',
